research/expert-sniper/runpod-worker/src/requirements.txt
runpod
python-dotenv
openai
